refactor : 길안내 시작시 기존 레디스 세션 삭제 후 생성, dto 및 파라미터 타입 변경
- 길안내 시작시 기존 세션 유무 확인하지 않고 기존 레디스 삭제후 생성하도록 수정.
- 길안내 관련 dto와 파라미터의 patientId, navigationId Long -> Integer로 수정

 S13P31A305-165

+++ backend/ongil/src/main/java/kr/co/ongil/domain/navigation/dto/request/StartNavigationRequest.java
@@ -10,7 +10,7 @@ public record StartNavigationRequest (
 
     @NotNull
     @Schema(description = "환자 ID", example = "123")
-    Long patientId,
+    Integer patientId,
 
     @NotNull
     @Schema(description = "출발지 정보")
